requirements.txt
asgiref==3.3.1
dj-database-url==0.5.0
Django==3.1.4
django-heroku==0.3.1
djangorestframework==3.12.2
mysqlclient==2.0.1
psycopg2==2.8.6
python-dateutil==2.8.1
pytz==2020.4
six==1.15.0
sqlparse==0.4.1
Unidecode==1.1.1
whitenoise==5.2.0
